+++ src/library/lib_mesh.hpp
@@ -6,6 +6,6 @@ namespace NiHu
 {
 
 NiHu::mesh<tmp::vector<NiHu::line_1_elem> > 
-create_line_1_circle_mesh(double r, int N);
+create_line_1_circle_mesh(double r, size_t N);
 
 }
